[Chore] as -> asChild 네이밍 수정, 공유하기 모달 trigger 위치 수정
